remove test container, update tests to run on host machine
