--- Dockerfile-tests
@@ -1,12 +1,0 @@
-FROM python:3.6
-
-RUN pip install --upgrade pip
-RUN pip install pipenv
-
-EXPOSE 8888
-
-COPY ./ /app
-WORKDIR /app/
-RUN pipenv install --system --dev
-
-ENV PYTHONPATH=/app
