+++ config/test.ini
@@ -4,4 +4,4 @@ TESTING =
 PORT = 80
 ATST_REDIRECT = http://localhost:8000/login
 ATST_PASSTHROUGH = http://localhost:8000/login-dev
-DATABASE_URI = postgres://atat:password@db/authnid
+DATABASE_URI = postgres://atat:password@localhost:5433/authnid
