--- docker-compose.override.yml
@@ -1,33 +1,0 @@
-version: '3'
-services:
-  backend:
-    build:
-      context: ./
-      args:
-        env: dev
-    volumes:
-      - './:/app'
-    environment:
-      - FLASK_APP=app/main.py
-      - FLASK_DEBUG=1
-      - 'RUN=flask run --host=0.0.0.0 --port=80'
-    # command: bash -c "while true; do sleep 1; done"
-    # command: bash -c "flask run --host=0.0.0.0 --port=80"
-  backend-tests:
-    build:
-      context: .
-      dockerfile: Dockerfile-tests
-      args:
-        env: dev
-    environment:
-      - SERVER_NAME=backend
-      # this sets the CA for the test requests
-      - REQUESTS_CA_BUNDLE=/app/ssl/certificate-authority/ca.crt
-      - FLASK_ENV
-      # - SEED=0
-    command: bash -c "while true; do sleep 1; done"
-    volumes:
-      - ./:/app
-      - ./ssl:/etc/ssl
-    ports:
-      - '8888:8888'
