+++ docker-compose.yml
@@ -14,3 +14,5 @@ services:
       - POSTGRES_DB=authnid
       - POSTGRES_USER=atat
       - POSTGRES_PASSWORD=password
+    ports:
+      - "5433:5432"
